[add] インターフェイスを分離
