+++ Assets/Scripts/Runtime/2.Application/OutGame/Scenario/IScenarioCompletionNotifier.cs
@@ -1,0 +1,10 @@
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace KillChord.Runtime.Application
+{
+    public interface IScenarioCompletionNotifier
+    {
+        ValueTask NotifyCompletedAsync(bool skipped, CancellationToken ct);
+    }
+}
