+++ Assets/Scripts/Runtime/2.Application/OutGame/Scenario/IScenarioCompletionNotifier.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 300b67f367c91704fbb43b5b6533f3ec
