+++ Assets/Scripts/Runtime/2.Application/OutGame/Scenario/IScenarioSettingsRepository.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 264b494c30d62d14388f90b08c0eab1f
