+++ Assets/Scripts/Runtime/2.Application/OutGame/Scenario/ITextAdvanceWaiter.cs
@@ -1,5 +1,4 @@
 
-using System;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -8,21 +7,5 @@ namespace KillChord.Runtime.Application
     public interface ITextAdvanceWaiter
     {
         public ValueTask WaitNextAsync(CancellationToken ct);
-    }
-
-    public interface IScenarioCompletionNotifier
-    {
-        ValueTask NotifyCompletedAsync(bool skipped, CancellationToken ct);
-    }
-
-    public interface IScenarioSettingsRepository
-    {
-        TimeSpan NormalTextCharInterval { get; }
-        TimeSpan FastForwardTextCharInterval { get; }
-        TimeSpan PausePollInterval { get; }
-        TimeSpan CloseDelayAfterComplete { get; }
-        bool SkipClosesImmediately { get; }
-        bool WaitForInputOnLastText { get; }
-        string DefaultScenarioId { get; }
     }
 }
