+++ Assets/Scripts/Runtime/3.Adaptor/OutGame/Scenario/IScenarioCompletionViewSink.cs
@@ -1,0 +1,7 @@
+namespace KillChord.Runtime.Adaptor
+{
+    public interface IScenarioCompletionViewSink
+    {
+        void SetScenarioCompleted(bool skipped);
+    }
+}
