+++ Assets/Scripts/Runtime/3.Adaptor/OutGame/Scenario/IScenarioCompletionViewSink.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 64ab7b4ea06fbdb418227ad8eea8bdea
