+++ Assets/Scripts/Runtime/3.Adaptor/OutGame/Scenario/ITextViewSink.cs
@@ -4,9 +4,4 @@ namespace KillChord.Runtime.Adaptor
     {
         void SetText(string message);
     }
-
-    public interface IScenarioCompletionViewSink
-    {
-        void SetScenarioCompleted(bool skipped);
-    }
 }
