+++ Assets/Scripts/Runtime/5.InfraStructure/OutGame/Scenario/InMemoryScenarioRepository.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: d9bd649a7a6e7654297168f47b49c8dd
