+++ Assets/Scripts/Runtime/5.InfraStructure/OutGame/Scenario/ScenarioSettingsRepository.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 522088294c6f12b45a902597b2774fec
